requirements.txt
django-phonenumber-field
phonenumbers

Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt
psycopg2>=2.8.6
reportlab>=4.0
django-filter>=23.1
python-dateutil
drf-spectacular>= 0.15.1,<0.16
drf-nested-routers>=0.93.4

celery>=5.1
redis>=4.9
pytest>=8.0
pytest-django>=4.8
pandas>=2.2